BugCommit: TimeField defaultValues, Location/Link not open by default
Added functions to convert the stored 12 hr time format into 24 hr time format which <input> parses, however, making it show up as a defaultValue and then being able to be overwritten by changing values seems difficult.

If the default value for Mode is 'In-Person', the 'location' field doesn't display by default, and only displays once the value changes away from 'In-Person' and back. Same thing happens with 'Remote' and 'Link'.